增加错误信息收集并上传服务器功能 使用lint检查代码

+++ app/src/main/java/com/example/jobbook/moment/widget/MomentDetailActivity.java
@@ -337,7 +337,7 @@ public class MomentDetailActivity extends Activity implements MomentDetailView, 
             mTitleBarLayout.setLayoutParams(new LinearLayout.LayoutParams(LinearLayout.LayoutParams.
                     MATCH_PARENT, 0, (mTitleBarHeight / newHeight) * 568));
             mRecyclerView.setLayoutParams(new LinearLayout.LayoutParams(LinearLayout.LayoutParams.
-                    MATCH_PARENT, 0, ((float) (newHeight - 2 * mTitleBarHeight) / newHeight) * 568));
+                    MATCH_PARENT, 0, ((newHeight - 2 * mTitleBarHeight) / newHeight) * 568));
             mInputLayout.setLayoutParams(new LinearLayout.LayoutParams(LinearLayout.LayoutParams.
                     MATCH_PARENT, 0, (mInputLayoutHeight / newHeight) * 568));
             mTitleBarLayout.invalidate();
